Refactor WebSocket manager to separate module
Moved the WebsocketManager class and its instantiation from background.js to a new infrastructure/socketManager.js module. Updated background.js to import and use wsManager from the new module, improving code organization and separation of concerns.
